+++ src/internal/syscall_ret.c
@@ -1,9 +1,9 @@
 #include <errno.h>
 #include "syscall.h"
 
-long __syscall_ret(unsigned long r)
+long long __syscall_ret(unsigned long long r)
 {
-	if (r > -4096UL) {
+	if (r > -4096ULL) {
 		errno = -r;
 		return -1;
 	}
