+++ src/thread/__syscall_cp.c
@@ -1,9 +1,9 @@
 #include "pthread_impl.h"
 #include "syscall.h"
 
-hidden long __syscall_cp_c();
+hidden long long __syscall_cp_c();
 
-static long sccp(syscall_arg_t nr,
+static long long sccp(syscall_arg_t nr,
                  syscall_arg_t u, syscall_arg_t v, syscall_arg_t w,
                  syscall_arg_t x, syscall_arg_t y, syscall_arg_t z)
 {
@@ -12,7 +12,7 @@ static long sccp(syscall_arg_t nr,
 
 weak_alias(sccp, __syscall_cp_c);
 
-long (__syscall_cp)(syscall_arg_t nr,
+long long (__syscall_cp)(syscall_arg_t nr,
                     syscall_arg_t u, syscall_arg_t v, syscall_arg_t w,
                     syscall_arg_t x, syscall_arg_t y, syscall_arg_t z)
 {
